fix tests, fix doc tests, fix imports in indentity-account
