+++ examples/account/multiple_identities.rs
@@ -11,7 +11,7 @@ use identity::account::AccountStorage;
 use identity::account::IdentitySetup;
 use identity::account::Result;
 use identity::iota::ExplorerUrl;
-use identity::iota::IotaDID;
+use identity::iota_core::IotaDID;
 
 #[tokio::main]
 async fn main() -> Result<()> {
