+++ examples/low-level-api/revoke_vc.rs
@@ -19,10 +19,10 @@ use identity::did::MethodScope;
 use identity::did::DID;
 use identity::iota::CredentialValidation;
 use identity::iota::ExplorerUrl;
-use identity::iota::IotaVerificationMethod;
 use identity::iota::Receipt;
 use identity::iota::Resolver;
 use identity::iota::Result;
+use identity::iota_core::IotaVerificationMethod;
 use identity::prelude::*;
 
 mod common;
