+++ identity-account/src/account/account.rs
@@ -10,15 +10,15 @@ use serde::Serialize;
 use identity_core::crypto::SetSignature;
 use identity_core::crypto::SignatureOptions;
 use identity_iota::chain::DocumentChain;
-use identity_iota::did::IotaDID;
-use identity_iota::diff::DiffMessage;
-use identity_iota::document::IotaDocument;
-use identity_iota::document::IotaVerificationMethod;
 use identity_iota::document::ResolvedIotaDocument;
 use identity_iota::tangle::Client;
-use identity_iota::tangle::MessageId;
-use identity_iota::tangle::MessageIdExt;
 use identity_iota::tangle::PublishType;
+use identity_iota_core::did::IotaDID;
+use identity_iota_core::diff::DiffMessage;
+use identity_iota_core::document::IotaDocument;
+use identity_iota_core::document::IotaVerificationMethod;
+use identity_iota_core::message::MessageId;
+use identity_iota_core::message::MessageIdExt;
 
 use crate::account::AccountBuilder;
 use crate::account::PublishOptions;
