+++ identity-account/src/storage/stronghold.rs
@@ -10,7 +10,7 @@ use identity_core::crypto::PrivateKey;
 use identity_core::crypto::PublicKey;
 use identity_did::did::DID;
 use identity_did::verification::MethodType;
-use identity_iota::did::IotaDID;
+use identity_iota_core::did::IotaDID;
 use iota_stronghold::Location;
 use iota_stronghold::SLIP10DeriveInput;
 use std::convert::TryFrom;
