+++ identity-account/src/tests/account.rs
@@ -12,13 +12,13 @@ use identity_core::crypto::SignatureOptions;
 use identity_did::utils::Queryable;
 use identity_did::verification::MethodScope;
 use identity_iota::chain::DocumentChain;
-use identity_iota::diff::DiffMessage;
-use identity_iota::document::IotaDocument;
 use identity_iota::tangle::Client;
 use identity_iota::tangle::ClientBuilder;
-use identity_iota::tangle::MessageId;
-use identity_iota::tangle::MessageIdExt;
-use identity_iota::tangle::Network;
+use identity_iota_core::diff::DiffMessage;
+use identity_iota_core::document::IotaDocument;
+use identity_iota_core::message::MessageId;
+use identity_iota_core::message::MessageIdExt;
+use identity_iota_core::types::Network;
 
 use crate::account::Account;
 use crate::account::AccountBuilder;
@@ -268,7 +268,9 @@ async fn test_account_publish_options_sign_with() -> Result<()> {
       .publish_with_options(PublishOptions::default().sign_with("non-existent-method"))
       .await
       .unwrap_err(),
-    Error::IotaError(identity_iota::Error::InvalidDoc(identity_did::Error::MethodNotFound))
+    Error::IotaCoreError(identity_iota_core::Error::InvalidDoc(
+      identity_did::Error::MethodNotFound
+    ))
   ));
 
   assert!(matches!(
@@ -276,7 +278,9 @@ async fn test_account_publish_options_sign_with() -> Result<()> {
       .publish_with_options(PublishOptions::default().sign_with(auth_method))
       .await
       .unwrap_err(),
-    Error::IotaError(identity_iota::Error::InvalidDoc(identity_did::Error::MethodNotFound))
+    Error::IotaCoreError(identity_iota_core::Error::InvalidDoc(
+      identity_did::Error::MethodNotFound
+    ))
   ));
 
   // TODO: Once implemented, add a merkle key collection method with capability invocation relationship and test for
