+++ identity-iota-core/src/did/macros.rs
@@ -11,7 +11,7 @@
 ///
 /// ```
 /// # use identity_did::did::DID;
-/// # use identity_iota::try_construct_did;
+/// # use identity_iota_core::try_construct_did;
 /// #
 /// let did = try_construct_did!(b"public-key")?;
 /// assert_eq!(did.as_str(), "did:iota:2xQiiGHDq5gCi1H7utY1ni7cf65fTay3G11S4xKp1vkS");
@@ -21,7 +21,7 @@
 ///   did.as_str(),
 ///   "did:iota:com:2xQiiGHDq5gCi1H7utY1ni7cf65fTay3G11S4xKp1vkS"
 /// );
-/// # Ok::<(), identity_iota::Error>(())
+/// # Ok::<(), identity_iota_core::Error>(())
 /// ```
 #[macro_export]
 macro_rules! try_construct_did {
