+++ identity-iota-core/src/message/mod.rs
@@ -5,5 +5,6 @@
 pub use iota_client::bee_message::MessageId;
 
 pub use self::traits::MessageIdExt;
+pub use self::traits::TangleRef;
 
 mod traits;
