--- identity-iota-core/src/tangle/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2020-2022 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-pub use self::traits::TangleRef;
-
-mod traits;
