+++ identity-iota/src/chain/document_chain.rs
@@ -171,7 +171,7 @@ mod test {
   use identity_iota_core::did::IotaDIDUrl;
   use identity_iota_core::document::IotaDocument;
   use identity_iota_core::document::IotaVerificationMethod;
-  use identity_iota_core::tangle::TangleRef;
+  use identity_iota_core::message::TangleRef;
 
   use crate::Error;
