+++ identity-iota/src/document/resolved_iota_document.rs
@@ -12,7 +12,7 @@ use identity_iota_core::diff::DiffMessage;
 use identity_iota_core::document::IotaDocument;
 use identity_iota_core::message::MessageId;
 use identity_iota_core::message::MessageIdExt;
-use identity_iota_core::tangle::TangleRef;
+use identity_iota_core::message::TangleRef;
 use serde::Deserialize;
 use serde::Serialize;
