+++ identity-iota/src/tangle/client.rs
@@ -12,7 +12,7 @@ use identity_iota_core::did::IotaDID;
 use identity_iota_core::diff::DiffMessage;
 use identity_iota_core::document::IotaDocument;
 use identity_iota_core::message::MessageId;
-use identity_iota_core::tangle::TangleRef;
+use identity_iota_core::message::TangleRef;
 use identity_iota_core::types::Network;
 
 use crate::chain::ChainHistory;
